backend/requirements.txt
colorama>=0.4.6
python-dotenv>=1.0.0
fastapi>=0.104.0
supabase>=2.0.0
python-multipart>=0.0.6
openai>=1.3.0
langchain>=0.0.335
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
instructor>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.24.0
prisma>=0.10.0
# Text extraction libraries
PyPDF2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.23
pdfplumber>=0.10.0